Migrate INTEGER/BIGINT columns storing u64 values to TEXT
u64 cannot be faithfully represented as SQL INTEGER/BIGINT (signed i64).  The
storage layer performs no math on these columns, so TEXT is safe.

Migrates amount, fee, expiry, and similar value columns to TEXT across all
databases (mint, wallet, auth SQLite/PostgreSQL, and Supabase).  Time columns
(created_time, updated_at, timestamp, etc.) are left as INTEGER/BIGINT since
they fit within i64 range.

Rust code updated: From<u64> for Value converts to Text, all .to_i64() calls
replaced with .to_u64(), and as i64 casts removed for value fields.

+++ crates/cdk-sql-common/src/mint/auth/mod.rs
@@ -182,7 +182,7 @@ where
                    "#,
             )?
             .bind("blinded_message", message.to_bytes().to_vec())
-            .bind("amount", u64::from(signature.amount) as i64)
+            .bind("amount", u64::from(signature.amount))
             .bind("keyset_id", signature.keyset_id.to_string())
             .bind("c", signature.c.to_bytes().to_vec())
             .execute(&self.inner)
