+++ crates/cdk-sql-common/src/mint/signatures.rs
@@ -105,7 +105,7 @@ where
                         "#,
                     )?
                     .bind("blinded_message", message.to_bytes().to_vec())
-                    .bind("amount", u64::from(signature.amount) as i64)
+                    .bind("amount", u64::from(signature.amount))
                     .bind("keyset_id", signature.keyset_id.to_string())
                     .bind("c", signature.c.to_bytes().to_vec())
                     .bind("quote_id", quote_id.as_ref().map(|q| q.to_string()))
@@ -130,7 +130,7 @@ where
                         DO UPDATE SET total_issued = keyset_amounts.total_issued + EXCLUDED.total_issued
                         "#,
                     )?
-                    .bind("amount", u64::from(signature.amount) as i64)
+                    .bind("amount", u64::from(signature.amount))
                     .bind("keyset_id", signature.keyset_id.to_string())
                     .execute(&self.inner)
                     .await?;
@@ -158,7 +158,7 @@ where
                             )
                             .bind("blinded_message", message.to_bytes().to_vec())
                             .bind("signed_time", current_time as i64)
-                            .bind("amount", u64::from(signature.amount) as i64)
+                            .bind("amount", u64::from(signature.amount))
                             .execute(&self.inner)
                             .await?;
 
@@ -170,7 +170,7 @@ where
                                 DO UPDATE SET total_issued = keyset_amounts.total_issued + EXCLUDED.total_issued
                                 "#,
                             )?
-                            .bind("amount", u64::from(signature.amount) as i64)
+                            .bind("amount", u64::from(signature.amount))
                             .bind("keyset_id", signature.keyset_id.to_string())
                             .execute(&self.inner)
                             .await?;
